Microservicios/services/media/requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.21
PyJWT==2.8.0
python-dotenv==1.0.0
xmltodict==0.13.0
dicttoxml==1.7.5
google-cloud-storage==2.14.0
